pyproject.toml
[tool.poetry]
name = "media-bridge"
version = "1.2.0"
description = "Python CLI that use yt-dlp to download videos and upload them to cloud"
authors = ["Jvillegasd <johnnyvillegaslrs@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
yt-dlp = "^2025.2.19"
pydantic = "^2.10.6"
google-api-python-client = "^2.165.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.0.0"
pre-commit = "^4.1.0"

[tool.poetry.scripts]
media-bridge = "media_bridge.core:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
src = ["src"]  # Source code directories
exclude = []

[tool.ruff.lint.pydocstyle]
convention = "pep257"

[tool.coverage.run]
branch = true
omit = [
    "tests/*",
    "tests/*",
    "*/tests/*",
    ".venv/*",
    "*tests*.py",
    "core.py",
    ]

[tool.coverage.report]
fail_under = 80
include_namespace_packages = true
skip_covered = true
exclude_also = [
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
    ]
